chore: Install ESLint and Prettier as dependencies
ESLint and Prettier should be installed as the project's development
dependencies. It helps contributors synchronize the version of each tool
and does not rely on arbitrary built into code editors.

Also, it enables running ESLint and Prettier within a CI environment
(e.g. GitHub Workflows), or possibly in a precommit script.