Do not feed robort a bloody nora if uneeded
From conversation here:
https://discord.com/channels/466605739838930955/864070665615179786/922699965392900176
If garbo for some reason doesn't want to go to the Sea, the robort will
end up with a wasted drink. If we can predict that we won't use it, then
we should skip it.
Note that this logic will get hairy if we want to go the Sea for more
complex things in the future.