Preliminary/Experimental marginal MPA estimation (#912)
* Set checkpoints at 25 and 75 adv left

* Experimenting with outlier detection

We use a Session instance if it exists, but if it doesn't (due to garbo being aborted and restarted) we then use the preferences if they were logged

* Only print details of outliers that are worth more than 1000meat

* Alternatively, only consider single drop items with value >= 1000 as outliers

* Shorter formatting for summary printout

* Improved outlier filtering

We estimate the frequency of outlier drops by comparing it against the entire session. If we have more than 1 instance of the drop across the entire session we reject classifying it as an outlier for now.

* Track outliers against barf turns only

We should only compare item drops against barf turns since our marginal turns are also used in barf mountain

* Experimenting with formatting for purely aesthetic reasons

* Revert "Track outliers against barf turns only"

This reverts commit 3fb5dc2d65838fdadf8ac5cb773f415d416cfee2.

* Revert "Experimenting with formatting for purely aesthetic reasons"

This reverts commit 2a72e67fda5bab379dd7bef8f97d235e2006fb16.

* Merging from main

* Don't overestimate MPA due to new EoD things like thesis

* Use a proper flag when checkpoints are set

* Use flag instead of comparing dates

* Significant refactor of marginal session

Track items over longer period since they are less subject to changing rates across barf turns

* Actually get the item diff between sessions

* Fix item tracking overestimate

* Don't track items all the way till the end

Since some EOD actions are not marginal, but intentionally put at our EOD turns

* Address scotch's concerns

* More merges

* Re-add porcelain candy dish

* revert

* Better accounting for marginal familiars

Subtract the excess value we obtain from running marginal familiars at the EoD. The remaining value is kept in items instead of being shifted over to meat, which is up for debate

* Fixed marginalFamiliarsExcessValue calc

* Added hardcoded outlier list

We expect to cap out the drops of certain items such as volcoino and extrovermectin, so they should never contribute to our marginal MPA

* reset the excess value to 0 after each update

* Account for jellyfish's outfitValue

* Add pokegro to outlier list since it has a cap