Safely diet even if fullness exceeds capacity after a Smol run
This targets #1543 where it's possible to have 20 / 15 fullness after
breaking the prism on a Shrunken Adventurer path run.

This aims to fix to the capacity calculation when evaluating diet
entries so that negative capacity isn't possible.

As a means of avoiding surprise loops, `consumeSafe` will now barf on
negative quantities.