Support bindlestocking crit drops
Unsure if the math will ever work out for this but worth calculating. Pretty much limited to only seal clubbers as otherwise it's difficult to crit when the weapon and off-hand slots are used.

The wiki drop rates I sourced from are probably slightly off but all that really matters here is what appears to be a 1% combined drop rate for fancy chocolate and jawbruisers